+++ cookbook/custom-plugin-decorators/update-example-dates/plugin.js
@@ -1,0 +1,12 @@
+import updateExampleDates from './decorator.js';
+
+export default function plugin() {
+  return {
+    id: 'dates-plugin',
+    decorators: {
+      oas3: {
+        'update-example-dates': updateExampleDates,
+      },
+    },
+  };
+}