+++ cookbook/custom-plugins/sorting/sort-enums.js
@@ -1,0 +1,12 @@
+export default function SortEnumsAlphabetically() {
+  console.log('re-ordering enums: alphabetical');
+  return {
+    Schema: {
+      leave(target) {
+        if (target.enum) {
+          target.enum.sort();
+        }
+      },
+    },
+  };
+}