+++ packages/respect-core/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @redocly/respect-core
 
+## 2.41.0
+
+### Patch Changes
+
+- Fixed an issue in `respect` where the execution of parent workflow's steps didn't halt after a step that referenced another workflow had failed.
+- Updated @redocly/openapi-core to v2.41.0.
+
+## 2.40.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.40.0.
+
+## 2.39.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.39.0.
+
+## 2.38.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.38.0.
+
 ## 2.37.0
 
 ### Patch Changes